started work on video fading
added a 3d model for chair, and worked on smoothly fading the video using alpha. smoothing mostly working (only issue is when activation / deactivation is interrupted). however, there are still issues, with the shader namely, that video doesn't actually work, and it also has some issues interacting with the vr goggles (only shows up in 1 eye for some reason)