added the ui into the "objects to deactive" in Obfuscation Manager
title, basically it was giving a warning last time, but it really is just because it is deactivated the coroutine for it fading in fails to start. It doesn't look bad though, and with juggling two seperate coroutine systems, I feel like it is more trouble than it is worth trying to figure out how to 'fix' it when it works fine now.